+++ src/rust/src/lib.rs
@@ -1,6 +1,7 @@
 use extendr_api::prelude::*;
 
 mod band_fetch;
+mod cell_mask;
 mod cell_raw;
 mod error;
 mod geo;